backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
mysqlclient==2.2.0
redis==4.6.0
celery==5.3.1
django-redis==5.3.0
Pillow==10.0.0 